Reorganize `subpageas/profile` in preparation to support fixed games
