+++ src/pages/profile.jsx
@@ -7,8 +7,7 @@ import Layout from '../components/layout'
 import { ButtonLink } from '../components/links'
 import Loading from '../components/loading'
 import { loadProfiles, ProfilePropType, saveProfiles } from '../helper/profiles'
-import PlayerDetails from '../subpages/profile/playerDetails'
-import GameHistory from '../subpages/profile/gameHistory'
+import ProfileContent from '../subpages/profile/content'
 
 const activeProfileStorageKey = 'shengji-helper-active'
 const redirectTo = (window, dest) => window.location.replace(dest)
@@ -83,11 +82,7 @@ const Profile = ({ location }) => {
 
     return (<Layout header={<ProfileHeader profile={profileList && profileList[uuid]} />}>
         {(profileList && profileList[uuid]) ?
-            <>
-                <PlayerDetails profile={profileList[uuid]} onUpdate={handleProfileUpdate} mb={2} />
-                <GameHistory profile={profileList[uuid]} />
-            </> :
-            <Loading />}
+            <ProfileContent profile={profileList[uuid]} onUpdate={handleProfileUpdate} /> : <Loading />}
     </Layout>)
 }
